Rename file to 'metric_helpers' to emphasize that they are not everything
